refactor underway
