+++ simple.cabal
@@ -92,6 +92,8 @@ library
     ExtractField,
     OpcodeToForm,
     Types,
+    Machine,
+    RegFiles,
     Util
   default-language: Haskell2010
