+++ src/Main.hs
@@ -7,42 +7,31 @@
 module Main where
 
 import Clash.Prelude
-import qualified Prelude as P
-import Types(Pc)
+import Machine(Machine(..),POWER_CPU(..))
+-- import qualified Prelude as P
 
-import Util(powerIndex32, powerIndex64)
-import OpcodeToForm(opcodeToForm)
-
-
-getRsSlice insn = slice (powerIndex32 @0) (powerIndex32 @5) insn
-val32 = 0x1C_34_56_78 :: Unsigned 32
-
-testValSlice :: Unsigned 6
-testValSlice = unpack $ getRsSlice val32
 
 
 -- currently memory is only 8 words deep
 memInit :: [Unsigned 32]
 memInit = [
-  0x1,
-  2,
-  3,
-  4,
-  5]
+  0x0000a03c,
+  0x3000a5e8,
+  0x1a002038,
+  0x18002598,
+  0x10002588,
+  0x01002170,
+  0xf8ff8141,
+  0x08002588,
+  0x01002138,
+  0x00002598,
+  0xe8ffff4b,
+  0x00000060,
+  0x002000c0,
+  0x00000000
+  ]
 
-type GPR = Vec 32 (Unsigned 64) -- General Purpose Registers page 10
-type Mem n = Vec n (Unsigned 32)
 
-
-data POWER_CPU = POWER_CPU
-  { pc :: Pc,
-    gpr :: GPR
-  } deriving (Show)
-
-data Machine = Machine 
-  { cpu :: POWER_CPU,
-    mem :: Mem 1024
-  } deriving (Show)
 
 -- Placeholder function that currently just increments
 -- the first entry in memory by 1
@@ -58,8 +47,8 @@ machine' machine@(Machine { cpu = POWER_CPU { pc = pc }, mem = mem }) =
 
 main :: IO ()
 main = do
-  putStrLn final_msg
+  putStrLn msg
   where
     msg :: String = "Hello from Main!"
-    strVal :: String = show testValSlice
-    final_msg :: String = msg P.++ strVal
+    -- strVal :: String = show testValSlice
+    -- final_msg :: String = msg P.++ strVal
