+++ src/Util.hs
@@ -5,7 +5,8 @@
 
 module Util(
   powerIndex32,
-  powerIndex64) where
+  powerIndex64,
+  endianSwap) where
 
 import Clash.Prelude
 
@@ -28,3 +29,14 @@ mkValidIndex64 = ValidIndex63 $ SNat @n
 powerIndex64 :: forall n. (KnownNat n, 0 <= n, n <= 63) => SNat (63 - n)
 powerIndex64 = case mkValidIndex64 @n of
   ValidIndex63 _ -> SNat @(63 - n)
+
+endianSwap :: Unsigned 32 -> Unsigned 32
+endianSwap x = (byte0 `shiftL` 24) .|.
+               (byte1 `shiftL` 16) .|.
+               (byte2 `shiftL` 8)  .|.
+               byte3
+  where
+    byte0 = (x .&. 0x000000FF)
+    byte1 = (x .&. 0x0000FF00) `shiftR` 8
+    byte2 = (x .&. 0x00FF0000) `shiftR` 16
+    byte3 = (x .&. 0xFF000000) `shiftR` 24
